pyproject.toml
[project]
name = "llm-wiki"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "edgeparse>=0.2.3",
    "langchain>=1.2.17",
    "langchain-core>=1.3.2",
    "langchain-ollama>=1.1.0",
    "langchain-openai>=1.1.0",
    "langgraph>=1.1.10",
    "mammoth>=1.11.0",
    "markdownify>=1.2.0",
    "python-dotenv>=1.2.1",
    "streamlit>=1.57.0",
]
